backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Google Cloud
google-cloud-firestore==2.14.0
google-generativeai==0.3.2

# GitHub
PyGithub==2.1.1

# Chatwork
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
